2019-12-22 Roman (stable) - !!:FIX:DB.LibsDB.vcaBase: Main.{terminator,ElCadr}: Revised during the documenting.     FIX:Main.calendar: Fixed for month start in Sunday (0) in the weeks form and background of the disabled items.     FIX:Fixed for the message "reversed" translation. - FIX:DB.LibsDB.OscadaLibs: DevLib.{1W_DS9097,W_DS9097U}: Fixed for the device family detection as the template 0x7F,     detected on Last DS2413 with the first byte 0xBA. - DB: The SQL-requesting dialog appended for the time measurement. - DAQ: TDAQS::rdRestDtTm() expanded from the float to double type. - FIX:SYS: The redundancy task fixed moving the declaration "XMLNode req("CntrReqs");" up to the same request,     due to the possibility of back reading wrong responses. - FIX:Transport: Fixed in the transports copying for additional parameters, by call load_(). - FIX:DAQ.JavaLikeCalc: Appended for disabling controller objects at changing the function. - !:UI.Vision: Appended for an option of the module to drop widget styles to the common one for some specific widgets in the runtime. - FIX:UI.WebVision: The views "Data" of the primitive "FormEl" fixed in the calendar for month start in Sunday (0)     in the weeks form and background of the disabled items.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2655 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tdaqs.h
@@ -1,7 +1,7 @@
 
 //OpenSCADA file: tdaqs.h
 /***************************************************************************
- *   Copyright (C) 2003-2018 by Roman Savochenko, <rom_as@oscada.org>      *
+ *   Copyright (C) 2003-2019 by Roman Savochenko, <rom_as@oscada.org>      *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -73,8 +73,8 @@ class TDAQS : public TSubSYS
 
 	// Redundancy
 	bool rdProcess( XMLNode *reqSt = NULL );
-	float rdRestDtTm( )		{ return mRdRestDtTm; }
-	void setRdRestDtTm( float vl )	{ mRdRestDtTm = vmin(24,vmax(0.01,vl)); modif(); }
+	double rdRestDtTm( )		{ return mRdRestDtTm; }
+	void setRdRestDtTm( double vl )	{ mRdRestDtTm = vmin(24,vmax(0.01,vl)); modif(); }
 	void rdActCntrList( vector<string> &ls, bool isRun = false );
 	string rdStRequest( const string &cntr, XMLNode &req, const string &prevSt = "", bool toRun = true );
 
@@ -102,7 +102,7 @@ class TDAQS : public TSubSYS
 	int	mTmpLib;
 
 	ResRW	mRdRes;
-	float	mRdRestDtTm;			//Depth time of the redundant history restore in hours
+	double	mRdRestDtTm;			//Depth time of the redundant history restore in hours
 	map<string, map<string,bool> > mRdCntr;
 };
